accounting_storage/mysql - Fix uint16_t truncation of node event state
Node event state is written to the DB as a uint32_t but read back into a
uint16_t, silently dropping all flag bits at position 16 and above.

This isn't actually reachable from current code.

Fixes regression from commit 845ff7d46a.

+++ src/plugins/accounting_storage/mysql/as_mysql_rollup.c
@@ -937,7 +937,7 @@ static local_cluster_usage_t *_setup_cluster_usage(mysql_conn_t *mysql_conn,
 	while ((row = mysql_fetch_row(result))) {
 		time_t row_start = slurm_atoul(row[EVENT_REQ_START]);
 		time_t row_end = slurm_atoul(row[EVENT_REQ_END]);
-		uint16_t state = slurm_atoul(row[EVENT_REQ_STATE]);
+		uint32_t state = slurm_atoul(row[EVENT_REQ_STATE]);
 		time_t local_start, local_end;
 		int seconds, resv_seconds;
